+++ src/scripts/domWidget.test.ts
@@ -3,8 +3,13 @@ import { describe, expect, test, vi } from 'vitest'
 
 import type { LGraph } from '@/lib/litegraph/src/litegraph'
 import { LGraphNode } from '@/lib/litegraph/src/litegraph'
-import { ComponentWidgetImpl, DOMWidgetImpl } from '@/scripts/domWidget'
-import { addWidget, isComponentWidget, isDOMWidget } from '@/scripts/domWidget'
+import {
+  addWidget,
+  ComponentWidgetImpl,
+  DOMWidgetImpl,
+  isComponentWidget,
+  isDOMWidget
+} from '@/scripts/domWidget'
 
 const { registerWidget, unregisterWidget } = vi.hoisted(() => ({
   registerWidget: vi.fn(),
@@ -33,13 +38,11 @@ describe('DOMWidget Y Position Preservation', () => {
       options: {}
     })
 
-    // Set a specific Y position
     originalWidget.y = 66
 
     const newNode = new LGraphNode('new-node')
     const clonedWidget = originalWidget.createCopyForNode(newNode)
 
-    // Verify Y position is preserved
     expect(clonedWidget.y).toBe(66)
     expect(clonedWidget.node).toBe(newNode)
     expect(clonedWidget.name).toBe('test-widget')
@@ -57,13 +60,11 @@ describe('DOMWidget Y Position Preservation', () => {
       options: {}
     })
 
-    // Set a specific Y position
     originalWidget.y = 42
 
     const newNode = new LGraphNode('new-node')
     const clonedWidget = originalWidget.createCopyForNode(newNode)
 
-    // Verify Y position is preserved
     expect(clonedWidget.y).toBe(42)
     expect(clonedWidget.node).toBe(newNode)
     expect(clonedWidget.element).toBe(mockElement)
@@ -80,11 +81,9 @@ describe('DOMWidget Y Position Preservation', () => {
       options: {}
     })
 
-    // Don't explicitly set Y (should be 0 by default)
     const newNode = new LGraphNode('new-node')
     const clonedWidget = originalWidget.createCopyForNode(newNode)
 
-    // Verify Y position is preserved (should be 0)
     expect(clonedWidget.y).toBe(0)
   })
 })
